% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_monitor_cluster.R
\name{crew_class_monitor_cluster}
\alias{crew_class_monitor_cluster}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Abstract cluster monitor class}
\description{
Abstract cluster monitor \code{R6} class
}
\details{
See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.
}
\seealso{
Other cluster: 
\code{\link{crew_class_launcher_cluster}},
\code{\link{crew_launcher_cluster}()},
\code{\link{crew_monitor_cluster}()}
}
\concept{cluster}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{verbose}}{See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.}

\item{\code{command_list}}{See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.}

\item{\code{command_terminate}}{See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_monitor_cluster-new}{\code{crew_class_monitor_cluster$new()}}
\item \href{#method-crew_class_monitor_cluster-validate}{\code{crew_class_monitor_cluster$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_cluster-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_cluster-new}{}}}
\subsection{Method \code{new()}}{
Abstract cluster monitor constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_cluster$new(
  verbose = NULL,
  command_list = NULL,
  command_terminate = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.}

\item{\code{command_list}}{See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.}

\item{\code{command_terminate}}{See \code{\link[=crew_monitor_cluster]{crew_monitor_cluster()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an abstract cluster monitor object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_cluster-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_cluster-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the monitor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_cluster$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
