% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_monitor_sge.R
\name{crew_monitor_sge}
\alias{crew_monitor_sge}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Create a SGE monitor object.}
\usage{
crew_monitor_sge(
  verbose = TRUE,
  command_list = as.character(Sys.which("qstat")),
  command_terminate = as.character(Sys.which("qdel"))
)
}
\arguments{
\item{verbose}{Deprecated. Use \code{options_cluster} instead.}

\item{command_list}{Character of length 1,
file path to the executable to list jobs.}

\item{command_terminate}{Deprecated. Use \code{options_cluster} instead.}
}
\description{
Create an \code{R6} object to monitor SGE cluster jobs.
}
\seealso{
Other sge: 
\code{\link{crew_class_launcher_sge}},
\code{\link{crew_class_monitor_sge}},
\code{\link{crew_controller_sge}()},
\code{\link{crew_launcher_sge}()},
\code{\link{crew_options_sge}()}
}
\concept{sge}
