% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_monitor_slurm.R
\name{crew_monitor_slurm}
\alias{crew_monitor_slurm}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Create a SLURM monitor object.}
\usage{
crew_monitor_slurm(
  verbose = TRUE,
  command_list = as.character(Sys.which("squeue")),
  command_terminate = as.character(Sys.which("scancel"))
)
}
\arguments{
\item{verbose}{Deprecated. Use \code{options_cluster} instead.}

\item{command_list}{Character of length 1,
file path to the executable to list jobs.}

\item{command_terminate}{Deprecated. Use \code{options_cluster} instead.}
}
\description{
Create an \code{R6} object to monitor SLURM cluster jobs.
}
\seealso{
Other slurm: 
\code{\link{crew_class_launcher_slurm}},
\code{\link{crew_class_monitor_slurm}},
\code{\link{crew_controller_slurm}()},
\code{\link{crew_launcher_slurm}()},
\code{\link{crew_options_slurm}()}
}
\concept{slurm}
