% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_package.R
\docType{package}
\name{crew.cluster-package}
\alias{crew.cluster-package}
\title{crew.cluster: crew launcher plugins for traditional high-performance
computing clusters}
\description{
In computationally demanding analysis projects,
statisticians and data scientists asynchronously
deploy long-running tasks to distributed systems,
ranging from traditional clusters to cloud services.
The \code{crew.cluster} package extends the
\href{https://github.com/shikokuchuo/mirai}{\code{mirai}}-powered
\href{https://wlandau.github.io}{\code{crew}} package with worker
launcher plugins for traditional
high-performance computing systems.
Inspiration also comes from packages
\href{https://github.com/shikokuchuo/mirai}{\code{mirai}},
\href{https://future.futureverse.org/}{\code{future}},
\href{https://mrc-ide.github.io/rrq/}{\code{rrq}},
\href{https://mschubert.github.io/clustermq/}{\code{clustermq}},
and \href{https://mllg.github.io/batchtools/}{\code{batchtools}}.
}
\section{Attribution}{

The template files at
\url{https://github.com/mschubert/clustermq/tree/master/inst}
informed the development of the \code{crew} launcher plugins in
\code{crew.cluster}, and we would like to thank
Michael Schubert for developing \code{clustermq} and releasing it under
the permissive Apache License 2.0.
See the \code{NOTICE} and \code{README.md} files in the \code{crew.cluster}
source code for additional attribution.
}

\concept{help}
