% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher_lsf.R
\name{crew_class_launcher_lsf}
\alias{crew_class_launcher_lsf}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} LSF launcher class}
\description{
\code{R6} class to launch and manage LSF workers.
}
\details{
See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.
}
\section{Attribution}{

The template files at
\url{https://github.com/mschubert/clustermq/tree/master/inst}
informed the development of the \code{crew} launcher plugins in
\code{crew.cluster}, and we would like to thank
Michael Schubert for developing \code{clustermq} and releasing it under
the permissive Apache License 2.0.
See the \code{NOTICE} and \code{README.md} files in the \code{crew.cluster}
source code for additional attribution.
}

\examples{

## ------------------------------------------------
## Method `crew_class_launcher_lsf$script`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_launcher_lsf(
  lsf_cwd = getwd(),
  lsf_log_output = "log_file_\%J.log",
  lsf_log_error = NULL,
  lsf_memory_gigabytes_limit = 4
)
launcher$script(name = "my_job_name")
}
}
\seealso{
Other launchers: 
\code{\link{crew_class_launcher_cluster}},
\code{\link{crew_class_launcher_pbs}},
\code{\link{crew_class_launcher_sge}},
\code{\link{crew_class_launcher_slurm}},
\code{\link{crew_launcher_cluster}()},
\code{\link{crew_launcher_lsf}()},
\code{\link{crew_launcher_pbs}()},
\code{\link{crew_launcher_sge}()},
\code{\link{crew_launcher_slurm}()}
}
\concept{launchers}
\section{Super classes}{
\code{\link[crew:crew_class_launcher]{crew::crew_class_launcher}} -> \code{\link[crew.cluster:crew_class_launcher_cluster]{crew.cluster::crew_class_launcher_cluster}} -> \code{crew_class_launcher_lsf}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lsf_cwd}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_log_output}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_log_error}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_memory_gigabytes_limit}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_memory_gigabytes_required}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_cores}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_launcher_lsf-new}{\code{crew_class_launcher_lsf$new()}}
\item \href{#method-crew_class_launcher_lsf-validate}{\code{crew_class_launcher_lsf$validate()}}
\item \href{#method-crew_class_launcher_lsf-script}{\code{crew_class_launcher_lsf$script()}}
\item \href{#method-crew_class_launcher_lsf-args_launch}{\code{crew_class_launcher_lsf$args_launch()}}
\item \href{#method-crew_class_launcher_lsf-args_terminate}{\code{crew_class_launcher_lsf$args_terminate()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="backlogged"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-backlogged'><code>crew::crew_class_launcher$backlogged()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="call"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-call'><code>crew::crew_class_launcher$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="done"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-done'><code>crew::crew_class_launcher$done()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="launch"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-launch'><code>crew::crew_class_launcher$launch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="resolved"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-resolved'><code>crew::crew_class_launcher$resolved()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="rotate"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-rotate'><code>crew::crew_class_launcher$rotate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="scale"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-scale'><code>crew::crew_class_launcher$scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="settings"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-settings'><code>crew::crew_class_launcher$settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="start"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-start'><code>crew::crew_class_launcher$start()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="tally"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-tally'><code>crew::crew_class_launcher$tally()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="terminate"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-terminate'><code>crew::crew_class_launcher$terminate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="throttle"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-throttle'><code>crew::crew_class_launcher$throttle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="unlaunched"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-unlaunched'><code>crew::crew_class_launcher$unlaunched()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew.cluster" data-topic="crew_class_launcher_cluster" data-id="launch_worker"><a href='../../crew.cluster/html/crew_class_launcher_cluster.html#method-crew_class_launcher_cluster-launch_worker'><code>crew.cluster::crew_class_launcher_cluster$launch_worker()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew.cluster" data-topic="crew_class_launcher_cluster" data-id="terminate_worker"><a href='../../crew.cluster/html/crew_class_launcher_cluster.html#method-crew_class_launcher_cluster-terminate_worker'><code>crew.cluster::crew_class_launcher_cluster$terminate_worker()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_lsf-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_lsf-new}{}}}
\subsection{Method \code{new()}}{
LSF launcher constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_lsf$new(
  name = NULL,
  seconds_interval = NULL,
  seconds_launch = NULL,
  seconds_idle = NULL,
  seconds_wall = NULL,
  seconds_exit = NULL,
  tasks_max = NULL,
  tasks_timers = NULL,
  reset_globals = NULL,
  reset_packages = NULL,
  reset_options = NULL,
  garbage_collection = NULL,
  verbose = NULL,
  command_submit = NULL,
  command_delete = NULL,
  script_directory = NULL,
  script_lines = NULL,
  lsf_cwd = NULL,
  lsf_log_output = NULL,
  lsf_log_error = NULL,
  lsf_memory_gigabytes_limit = NULL,
  lsf_memory_gigabytes_required = NULL,
  lsf_cores = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{seconds_interval}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{seconds_launch}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{seconds_idle}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{seconds_wall}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{seconds_exit}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{tasks_max}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{tasks_timers}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{reset_globals}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{reset_packages}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{reset_options}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{garbage_collection}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{verbose}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{command_submit}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{command_delete}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{script_directory}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{script_lines}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_cwd}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_log_output}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_log_error}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_memory_gigabytes_limit}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_memory_gigabytes_required}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}

\item{\code{lsf_cores}}{See \code{\link[=crew_launcher_lsf]{crew_launcher_lsf()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an LSF launcher object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_lsf-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_lsf-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_lsf$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly). Throws an error if a field is invalid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_lsf-script"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_lsf-script}{}}}
\subsection{Method \code{script()}}{
Generate the job script.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_lsf$script(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, name of the job. For inspection
purposes, you can supply a mock job name.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Includes everything except the worker-instance-specific
job name and the worker-instance-specific
call to \code{crew::crew_worker()}, both of which get inserted at
the bottom of the script at launch time.
}

\subsection{Returns}{
Character vector of the lines of the job script.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_launcher_lsf(
  lsf_cwd = getwd(),
  lsf_log_output = "log_file_\%J.log",
  lsf_log_error = NULL,
  lsf_memory_gigabytes_limit = 4
)
launcher$script(name = "my_job_name")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_lsf-args_launch"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_lsf-args_launch}{}}}
\subsection{Method \code{args_launch()}}{
Worker launch arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_lsf$args_launch(script)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{script}}{Character of length 1, path to the job script for
the scheduler.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of arguments to the command that
launches a worker.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_lsf-args_terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_lsf-args_terminate}{}}}
\subsection{Method \code{args_terminate()}}{
Termination arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_lsf$args_terminate(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, name of the job of the worker
on the scheduler.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of arguments to the command that
terminates a worker.
}
}
}
