% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher_slurm.R
\name{crew_class_launcher_slurm}
\alias{crew_class_launcher_slurm}
\title{SLURM launcher class}
\description{
\code{R6} class to launch and manage SLURM workers.
}
\details{
See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.
}
\examples{

## ------------------------------------------------
## Method `crew_class_launcher_slurm$script`
## ------------------------------------------------

launcher <- crew_launcher_slurm(
  slurm_log_output = "log_file_\%A.log",
  slurm_log_error = NULL,
  slurm_memory_megabytes_per_cpu = 4096
)
launcher$script(name = "my_job_name")
}
\seealso{
Other launchers: 
\code{\link{crew_class_launcher_cluster}},
\code{\link{crew_class_launcher_sge}},
\code{\link{crew_launcher_cluster}()},
\code{\link{crew_launcher_sge}()},
\code{\link{crew_launcher_slurm}()}
}
\concept{launchers}
\section{Super classes}{
\code{\link[crew:crew_class_launcher]{crew::crew_class_launcher}} -> \code{\link[crew.cluster:crew_class_launcher_cluster]{crew.cluster::crew_class_launcher_cluster}} -> \code{crew_class_launcher_slurm}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{slurm_log_output}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{slurm_log_error}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{slurm_memory_megabytes_per_cpu}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{slurm_cpus_per_task}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_launcher_slurm-new}{\code{crew_class_launcher_slurm$new()}}
\item \href{#method-crew_class_launcher_slurm-validate}{\code{crew_class_launcher_slurm$validate()}}
\item \href{#method-crew_class_launcher_slurm-script}{\code{crew_class_launcher_slurm$script()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="call"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-call'><code>crew::crew_class_launcher$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="expected"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-expected'><code>crew::crew_class_launcher$expected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="launch"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-launch'><code>crew::crew_class_launcher$launch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="launching"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-launching'><code>crew::crew_class_launcher$launching()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="settings"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-settings'><code>crew::crew_class_launcher$settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="start"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-start'><code>crew::crew_class_launcher$start()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="terminate"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-terminate'><code>crew::crew_class_launcher$terminate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew.cluster" data-topic="crew_class_launcher_cluster" data-id="launch_worker"><a href='../../crew.cluster/html/crew_class_launcher_cluster.html#method-crew_class_launcher_cluster-launch_worker'><code>crew.cluster::crew_class_launcher_cluster$launch_worker()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew.cluster" data-topic="crew_class_launcher_cluster" data-id="terminate_worker"><a href='../../crew.cluster/html/crew_class_launcher_cluster.html#method-crew_class_launcher_cluster-terminate_worker'><code>crew.cluster::crew_class_launcher_cluster$terminate_worker()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_slurm-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_slurm-new}{}}}
\subsection{Method \code{new()}}{
SLURM launcher constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_slurm$new(
  name = NULL,
  seconds_launch = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL,
  seconds_idle = NULL,
  seconds_wall = NULL,
  seconds_exit = NULL,
  tasks_max = NULL,
  tasks_timers = NULL,
  reset_globals = NULL,
  reset_packages = NULL,
  reset_options = NULL,
  garbage_collection = NULL,
  verbose = NULL,
  command_submit = NULL,
  command_delete = NULL,
  script_directory = NULL,
  script_lines = NULL,
  slurm_log_output = NULL,
  slurm_log_error = NULL,
  slurm_memory_megabytes_per_cpu = NULL,
  slurm_cpus_per_task = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{seconds_launch}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{seconds_interval}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{seconds_timeout}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{seconds_idle}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{seconds_wall}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{seconds_exit}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{tasks_max}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{tasks_timers}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{reset_globals}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{reset_packages}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{reset_options}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{garbage_collection}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{verbose}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{command_submit}}{See \code{\link[=crew_launcher_sge]{crew_launcher_sge()}}.}

\item{\code{command_delete}}{See \code{\link[=crew_launcher_sge]{crew_launcher_sge()}}.}

\item{\code{script_directory}}{See \code{\link[=crew_launcher_sge]{crew_launcher_sge()}}.}

\item{\code{script_lines}}{See \code{\link[=crew_launcher_sge]{crew_launcher_sge()}}.}

\item{\code{slurm_log_output}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{slurm_log_error}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{slurm_memory_megabytes_per_cpu}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}

\item{\code{slurm_cpus_per_task}}{See \code{\link[=crew_launcher_slurm]{crew_launcher_slurm()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an SLURM launcher object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_slurm-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_slurm-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_slurm$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly). Throws an error if a field is invalid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_slurm-script"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_slurm-script}{}}}
\subsection{Method \code{script()}}{
Generate the job script.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_slurm$script(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, name of the job. For inspection
purposes, you can supply a mock job name.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Includes everything except the worker-instance-specific
job name and the worker-instance-specific
call to \code{crew::crew_worker()}, both of which get inserted at
the bottom of the script at launch time.
}

\subsection{Returns}{
Character vector of the lines of the job script.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{launcher <- crew_launcher_slurm(
  slurm_log_output = "log_file_\%A.log",
  slurm_log_error = NULL,
  slurm_memory_megabytes_per_cpu = 4096
)
launcher$script(name = "my_job_name")
}
\if{html}{\out{</div>}}

}

}
}
