% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_monitor_cluster.R
\name{crew_monitor_cluster}
\alias{crew_monitor_cluster}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Create an abstract
cluster monitor object.}
\usage{
crew_monitor_cluster(
  verbose = TRUE,
  command_list = as.character(Sys.which("qstat")),
  command_terminate = as.character(Sys.which("qdel"))
)
}
\arguments{
\item{verbose}{Logical, whether to see console output and error messages
when submitting worker.}

\item{command_list}{Character of length 1,
file path to the executable to list jobs.}

\item{command_terminate}{Character of length 1,
file path to the executable to terminate a worker job.
Set to \code{""} to skip manually terminating the worker.
Unless there is an issue with the platform,
the job should still exit thanks to the NNG-powered network programming
capabilities of \code{mirai}. Still, if you set \code{command_terminate = ""},
you are assuming extra responsibility for manually monitoring
your jobs on the cluster and manually terminating jobs as appropriate.}
}
\description{
Create an abstract cluster monitor \code{R6} object.
}
\seealso{
Other cluster: 
\code{\link{crew_class_launcher_cluster}},
\code{\link{crew_class_monitor_cluster}},
\code{\link{crew_launcher_cluster}()}
}
\concept{cluster}
\keyword{internal}
