% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_options_slurm.R
\name{crew_options_slurm}
\alias{crew_options_slurm}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} SLURM options.}
\usage{
crew_options_slurm(
  verbose = FALSE,
  command_submit = as.character(Sys.which("sbatch")),
  command_terminate = NULL,
  script_directory = tempdir(),
  script_lines = character(0L),
  log_output = "/dev/null",
  log_error = "/dev/null",
  memory_gigabytes_required = NULL,
  memory_gigabytes_per_cpu = NULL,
  cpus_per_task = NULL,
  time_minutes = NULL,
  partition = NULL,
  n_tasks = 1
)
}
\arguments{
\item{verbose}{Logical, whether to see console output and error messages
when submitting worker.}

\item{command_submit}{Character of length 1,
file path to the executable to submit a worker job.}

\item{command_terminate}{Deprecated on 2025-08-26 in
\code{crew.cluster} version 0.3.8.9001. No longer needed.}

\item{script_directory}{Character of length 1, directory path to the
job scripts. Just before each job submission, a job script
is created in this folder. Script base names are unique to each
launcher and worker, and the launcher deletes the script when the
worker is manually terminated. \code{tempdir()} is the default, but it
might not work for some systems.
\code{tools::R_user_dir("crew.cluster", which = "cache")}
is another reasonable choice.}

\item{script_lines}{Optional character vector of additional lines to be
added to the job script just after the more common flags.
An example would be \code{script_lines = "module load R"} if your cluster
supports R through an environment module.}

\item{log_output}{Character of length 1, file pattern to control
the locations of the SLURM worker log files. By default, both standard
output and standard error go to the same file.
\code{log_output = "crew_log_\%A.txt"} translates to a line of
\verb{#SBATCH --output=crew_log_\%A.txt} in the SLURM job script,
where \verb{\%A} is replaced by the job ID of the worker.
The default is \verb{/dev/null} to omit these logs.
Set \code{log_output = NULL} to omit this line from the job script.}

\item{log_error}{Character of length 1, file pattern for standard
error. \code{log_error = "crew_log_\%A.txt"} translates to a line of
\verb{#SBATCH --error=crew_log_\%A.txt} in the SLURM job script,
where \verb{\%A} is replaced by the job ID of the worker.
The default is \verb{/dev/null} to omit these logs.
Set \code{log_error = NULL} to omit this line from the job script.}

\item{memory_gigabytes_required}{Positive numeric scalar,
total number of gigabytes of memory required per node.
\code{memory_gigabytes_required = 2.40123}
translates to a line of \verb{#SBATCH --mem=2041M}
in the SLURM job script.
\code{memory_gigabytes_required = NULL} omits this line.}

\item{memory_gigabytes_per_cpu}{Positive numeric scalar,
gigabytes of
memory required per CPU.
\code{memory_gigabytes_per_cpu = 2.40123}
translates to a line of \verb{#SBATCH --mem-per-cpu=2041M}
in the SLURM job script.
\code{memory_gigabytes_per_cpu = NULL} omits this line.}

\item{cpus_per_task}{Optional positive integer scalar,
number of CPUs for the worker.
\code{cpus_per_task = 4} translates
to a line of \verb{#SBATCH --cpus-per-task=4} in the SLURM job script.
\code{cpus_per_task = NULL} omits this line.}

\item{time_minutes}{Numeric scalar,
number of minutes to
designate as the wall time of \code{crew} each worker instance on the
SLURM cluster. \code{time_minutes = 60} translates to a line of
\verb{#SBATCH --time=60} in the SLURM job script. \code{time_minutes = NULL}
omits this line.}

\item{partition}{Character string,
name of the SLURM partition to
create workers on. \code{partition = "partition1,partition2"}
translates to a line of \verb{#SBATCH --partition=partition1,partition2}
in the SLURM job script. \code{partition = NULL}
omits this line.}

\item{n_tasks}{Numeric scalar, number of SLURM tasks to run
within the job.
\code{n_tasks = 1} translates to a line of \verb{#SBATCH --ntasks=1}
in the SLURM job script.
\code{n_tasks = 0} omits this line.}
}
\value{
A classed list of options.
}
\description{
Set options for SLURM job management.
}
\section{Retryable options}{

Retryable options are deprecated in \code{crew.cluster} as of
2025-01-27 (version \verb{0.3.4}).
}

\section{Attribution}{

The template files at
\url{https://github.com/mschubert/clustermq/tree/master/inst}
informed the development of the \code{crew} launcher plugins in
\code{crew.cluster}, and we would like to thank
Michael Schubert for developing \code{clustermq} and releasing it under
the permissive Apache License 2.0.
See the \code{NOTICE} and \code{README.md} files in the \code{crew.cluster}
source code for additional attribution.
}

\examples{
  crew_options_slurm()
}
\seealso{
Other slurm: 
\code{\link{crew_class_launcher_slurm}},
\code{\link{crew_class_monitor_slurm}},
\code{\link{crew_controller_slurm}()},
\code{\link{crew_launcher_slurm}()},
\code{\link{crew_monitor_slurm}()}
}
\concept{slurm}
